`@remotion/browser-studio`: Support keyframe operations
