+++ packages/studio-codemods/src/sequence-props/can-update-effect-props.ts
@@ -37,8 +37,10 @@ const staticStatus = (codeValue: unknown): CanUpdateSequencePropStatus => ({
 	codeValue,
 });
 
-const findEffectsAttr = (jsx: JSXOpeningElement): JSXAttribute | null => {
-	for (const attr of jsx.attributes) {
+export const findEffectsAttr = (
+	attributes: JSXOpeningElement['attributes'],
+): JSXAttribute | null => {
+	for (const attr of attributes) {
 		if (attr.type !== 'JSXAttribute') {
 			continue;
 		}
@@ -49,6 +51,37 @@ const findEffectsAttr = (jsx: JSXOpeningElement): JSXAttribute | null => {
 	}
 
 	return null;
+};
+
+export const findEffectCallExpression = ({
+	attr,
+	effectIndex,
+}: {
+	attr: JSXAttribute;
+	effectIndex: number;
+}):
+	| {kind: 'ok'; call: CallExpression; callee: string}
+	| {kind: 'error'; reason: 'not-found' | 'not-call-expression'} => {
+	if (!attr.value || attr.value.type !== 'JSXExpressionContainer') {
+		return {kind: 'error', reason: 'not-call-expression'};
+	}
+
+	const expr = attr.value.expression as Expression;
+	if (expr.type !== 'ArrayExpression') {
+		return {kind: 'error', reason: 'not-call-expression'};
+	}
+
+	const elements = enumerateEffectArrayElements(expr);
+	if (effectIndex < 0 || effectIndex >= elements.length) {
+		return {kind: 'error', reason: 'not-found'};
+	}
+
+	const target = elements[effectIndex];
+	if (target.kind !== 'call') {
+		return {kind: 'error', reason: 'not-call-expression'};
+	}
+
+	return {kind: 'ok', call: target.node, callee: target.callee};
 };
 
 const getEffectsArrayElements = (
@@ -218,7 +251,7 @@ export const computeEffectPropStatus = ({
 	keys: string[];
 	videoConfigValues: VideoConfigIdentifierValues;
 }): CanUpdateEffectPropsResponse => {
-	const attr = findEffectsAttr(jsx);
+	const attr = findEffectsAttr(jsx.attributes);
 	const elements = getEffectsArrayElements(attr);
 
 	if (!elements) {
